+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- MPV window title now shows the episode number & anime title by default instead of part of the episode URL
+
 ## [0.2.2] - 2025-04-13
 
 ### Fixed